pyproject.toml
[project]
name = "github-error-investigator"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "openai-agents>=0.2.0",
    "pydantic>=2.9.0",
    "python-dotenv>=1.0.1",
    "requests>=2.32.0",
    "streamlit>=1.38.0",
    "uvicorn[standard]>=0.30.0",
]
